[BE] refact : IntakeReq - amount 자료형

+++ back/src/main/java/com/d101/back/entity/Intake.java
@@ -15,6 +15,6 @@ public class Intake {
 	@EmbeddedId
 	private FoodMealKey key;
 	
-	private int amount;
+	private Double amount;
 	
 }
